Edit first section html and css
I've editted first's section html and style this part. It looks like
it works. Circa same design and layout. I solved the problem with
shrinking. So now, when the page is to tight, the picture wrap to a
new line.

I tried solve the wrap problem in header. But there is some problem,
so it works, but no in the way I would like to.